fix findings
