+++ apps/sensenet/src/components/dialogs/reference-content-list/reference-list.tsx
@@ -1,7 +1,9 @@
+import { reactControlMapper } from '@sensenet/controls-react'
 import { GenericContent } from '@sensenet/default-content-types'
-import { useRepository } from '@sensenet/hooks-react'
+import { useLogger, useRepository } from '@sensenet/hooks-react'
 import {
   Avatar,
+  Button,
   IconButton,
   List,
   ListItem,
@@ -9,67 +11,141 @@ import {
   ListItemSecondaryAction,
   ListItemText,
 } from '@material-ui/core'
+import AddIcon from '@material-ui/icons/Add'
 import ClearIcon from '@material-ui/icons/Clear'
-import React from 'react'
+import React, { createElement, ReactNode, useState } from 'react'
 import { useGlobalStyles } from '../../../globalStyles'
+import { useLocalization, usePagination } from '../../../hooks'
 import { Icon } from '../../Icon'
 
 type ReferenceListProps = {
-  pagination: {
-    render: () => JSX.Element | null
-    currentItems: GenericContent[]
-    setCurrentPage: React.Dispatch<React.SetStateAction<number>>
-  }
   canEdit: boolean
-  references: GenericContent[]
-  setReferences: (refs: GenericContent[]) => void
   parent: GenericContent
   fieldName: string
+  items: GenericContent[]
+  renderButton?: (newRef: GenericContent | undefined) => ReactNode
+  formStyle?: React.CSSProperties
+  listStyle?: React.CSSProperties
 }
 
 export const ReferenceList: React.FC<ReferenceListProps> = (props) => {
   const globalClasses = useGlobalStyles()
   const repository = useRepository()
+  const controlMapper = reactControlMapper(repository)
+  const logger = useLogger('ReferenceContentList')
+  const localization = useLocalization()
+  const schema = controlMapper.getFullSchemaForContentType(props.parent.Type, 'new')
+  const field = schema.fieldMappings.find((fieldMap) => fieldMap.fieldSettings.Name === props.fieldName)
+  const [references, setReferences] = useState(props.items)
+  const [newReference, setNewReference] = useState<GenericContent>()
+  const [requestClearToken, setRequestClearToken] = useState<number>()
+  const pagination = usePagination({ items: references, color: 'primary' })
+
+  const fieldControl = createElement(
+    controlMapper.getControlForContentField(props.parent.Type, props.fieldName, 'new'),
+    {
+      repository,
+      settings: field!.fieldSettings,
+      content: props.parent,
+      actionName: 'new',
+      fieldOnChange: (_, value: GenericContent) => {
+        setNewReference(value)
+      },
+      triggerClear: requestClearToken,
+    },
+  )
+
+  const handleAddMembers = async (event: React.FormEvent<HTMLFormElement>) => {
+    event.preventDefault()
+
+    if (!references.find((reference) => reference.Id === newReference?.Id)) {
+      const newReferences = [newReference!, ...references]
+      await repository.patch({
+        idOrPath: props.parent.Id,
+        content: { [props.fieldName]: newReferences.map((ref) => ref.Id) },
+        forceRefresh: true,
+      })
+      setReferences(newReferences)
+    } else {
+      logger.error({
+        message: localization.referenceContentListDialog.errorAlreadyInList,
+        data: {
+          relatedContent: newReference,
+        },
+      })
+    }
+    setNewReference(undefined)
+    setRequestClearToken(new Date().getUTCMilliseconds())
+    pagination.setCurrentPage(1)
+  }
 
   return (
     <>
-      {props.pagination.currentItems.length ? (
-        <>
-          <List>
-            {props.pagination.currentItems.map((item) => (
-              <ListItem key={item.Id}>
-                <ListItemAvatar>
-                  <Avatar>
-                    <Icon item={item} />
-                  </Avatar>
-                </ListItemAvatar>
-                <ListItemText primary={item.DisplayName} />
-                {props.canEdit && (
-                  <ListItemSecondaryAction>
-                    <IconButton
-                      edge="end"
-                      aria-label="delete"
-                      onClick={async () => {
-                        const remainedReferences = props.references.filter((content) => item.Id !== content.Id)
-                        await repository.patch({
-                          idOrPath: props.parent.Id,
-                          content: { [props.fieldName]: remainedReferences.map((content) => content.Id) },
-                          forceRefresh: true,
-                        })
-                        props.setReferences(remainedReferences)
-                      }}>
-                      <ClearIcon />
-                    </IconButton>
-                  </ListItemSecondaryAction>
-                )}
-              </ListItem>
-            ))}
-          </List>
-          <div className={globalClasses.centeredHorizontal}>{props.pagination.render()}</div>
-        </>
-      ) : (
-        <div style={{ margin: '1rem 0' }}>No items</div>
+      {props.canEdit && (
+        <form
+          className={globalClasses.centeredVertical}
+          style={{
+            ...props.formStyle,
+          }}
+          onSubmit={handleAddMembers}>
+          {fieldControl}
+          {props.renderButton ? (
+            props.renderButton(newReference)
+          ) : (
+            <Button
+              variant="contained"
+              color="primary"
+              startIcon={<AddIcon />}
+              style={{ marginLeft: '0.5rem' }}
+              disabled={!newReference}
+              type="submit">
+              Add
+            </Button>
+          )}
+        </form>
       )}
+      <div
+        style={{
+          ...props.listStyle,
+        }}>
+        {pagination.currentItems.length ? (
+          <>
+            <List>
+              {pagination.currentItems.map((item) => (
+                <ListItem key={item.Id}>
+                  <ListItemAvatar>
+                    <Avatar>
+                      <Icon item={item} />
+                    </Avatar>
+                  </ListItemAvatar>
+                  <ListItemText primary={item.DisplayName} />
+                  {props.canEdit && (
+                    <ListItemSecondaryAction>
+                      <IconButton
+                        edge="end"
+                        aria-label="delete"
+                        onClick={async () => {
+                          const remainedReferences = references.filter((content) => item.Id !== content.Id)
+                          await repository.patch({
+                            idOrPath: props.parent.Id,
+                            content: { [props.fieldName]: remainedReferences.map((content) => content.Id) },
+                            forceRefresh: true,
+                          })
+                          setReferences(remainedReferences)
+                        }}>
+                        <ClearIcon />
+                      </IconButton>
+                    </ListItemSecondaryAction>
+                  )}
+                </ListItem>
+              ))}
+            </List>
+            <div className={globalClasses.centeredHorizontal}>{pagination.render()}</div>
+          </>
+        ) : (
+          <div style={{ margin: '1rem 0' }}>No items</div>
+        )}
+      </div>
     </>
   )
 }
